requirements.txt
Pillow
flask
requests
beautifulsoup4
